tcp: Add tcptest.py script
To test the TCP stack, add a simple script what will use the
echo-server sample in the Zephyr distribution to send some data, see if
the same data gets sent back, or if there were timeouts or different
data is received.

This does not perform a test at the protocol level, checking if the TCP
flags make sense.  Its only purpose is to assess the stability of the
new IP stack.

Change-Id: If55b2db52230ff6f84b644c0c1f677e0a1544c80
Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>