tinydtls: Fix compilation errors
ECC test files are not required for net-tools and which are causing
compilation errors.

/usr/bin/ld: ecc_test.o: relocation R_X86_64_32 against `.rodata'
can not be used when making a shared object; recompile with -fPIC
/usr/bin/ld: test_helper.o: relocation R_X86_64_32 against
`.rodata.str1.1' can not be used when making a shared object;
recompile with -fPIC

Also failed to build with -fPIC option. So just commenting those
programs.

Change-Id: I7f2a49737523b6f0fe343e134a877040dc374027
Signed-off-by: Ravi kumar Veeramally <ravikumar.veeramally@linux.intel.com>