tunslip6: strengthen and defer config to external parties
When started by a script, sometimes QEMU is not yet ready to receive
and tunslip6 stops; so added a retry in the serial_to_tun() path.

As part of the same effort, if the ipaddr is not specified, then the
interface won't be configured and that will be left to someone else
to do (in our case, the caller).

Change-Id: I23e0a12f80d87df32d477a1311f2bdedd90974d0
Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>