tunslip6: reload slip without restarting tap
This makes testing compatible with programs which open a raw
socket on an interface and can't be told they need to reopen
their socket when the tap0 interface is recreated.

slipfd and inslip are reopened after a wait period when an
error occurs reading from inslip. This makes it so we don't
have to bring down the tap interface we can continue with the
new slip and the old tap in case other programs are using it.

Signed-off-by: John Andersen <john.s.andersen@intel.com>