tunslip6: Flush neighbor cache when qemu restarts
If the qemu is restarted, the tunslip6 notices it and reopens the
interface. But we should flush the neighbor cache as the MAC addresses
of the qemu zephyr instance will be changed but the IP addresses will
be the same. If we do not flush, then Linux will take several seconds
before connectivity is restored after qemu restart.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>