net-setup: zeth-vlan.conf: Add IP addresses to zeth
Earlier the zeth interface did not had any IP address in it.
This is unnecessary and confusing so add the IPv4 and IPv6
addresses to that interface too. After this this config
creates three network interfaces, the zeth is normal interface
without VLAN, and then there are two VLAN interfaces called
vlan.100 and vlan.200 that run on top of zeth.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>