docker: Add mosquitto MQTT server to net-tools image
In addition to the already existing net tools, clone mosquitto
MQTT server from git, compile and install it. When running the
git clone, add any possible proxies as environment variables in
both RUN commands, e.g.:

   RUN https_proxy=proxy.example.net:112 git clone https://...

When mosquitto compilation is finished, install it in the Docker
image under /usr/local/ and remove the source code and build
directory. Mosquitto also needs a user id 'mosquitto'.

As multiple FROM directives are not that helpful, count on the
gcc image to contain necessary tools and build environment.

Signed-off-by: Patrik Flykt <patrik.flykt@linux.intel.com>