echo-client: Add VLAN specific classify packet option
Add -c option which tells the echo-client to set the socket
priority so that the system can set the VLAN priority of the
sent packet. This requires that net-setup.sh is started with
VLAN support and the system is set to map socket priority to
VLAN priority. One such config file is zeth-vlan-priority.conf
file found in this project.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>